change unittest config file
