+++ restdemo/tests/test_login.py
@@ -1,26 +1,9 @@
-import unittest
 import json
 
-from restdemo import create_app, db
+from restdemo.tests.base import TestBase
 
 
-class TestLogin(unittest.TestCase):
-
-    def setUp(self):
-        self.app = create_app(config_name='testing')
-        self.client = self.app.test_client
-        self.user_data = {
-            'username': 'test',
-            'password': 'test123',
-            'email': 'test123@test.com'
-        }
-        with self.app.app_context():
-            db.create_all()
-    
-    def tearDown(self):
-        with self.app.app_context():
-            db.session.remove()
-            db.drop_all()
+class TestLogin(TestBase):
 
     def test_login(self):
         url = '/user/{}'.format(self.user_data['username'])
